added url to links
This does the basic function of locating links and returning them as a list of strings.

in matlab it would have been its own separate function, but not sure how you want to organise this project (modules of several functions, or modules of singles functions, etc...) , and how to deal with imports within functions, etc...

I can finalise once we've discussed.